Update record.ru.txt
Hello developers! I understand that it is very difficult for you to translate the game into Russian. Since I am Russian myself, I managed to translate the game to normal. I'm glad to help you, because SD Tank War is a beginner game and it needs community support. Thanks for the great game!